JWT implemented to login and create a user generating a jwt.

+++ src/main/java/com/unir/Usuarios/service/AmigosService.java
@@ -2,6 +2,7 @@ package com.unir.Usuarios.service;
 
 import com.unir.Usuarios.data.AmigosJpaRepository;
 import com.unir.Usuarios.model.db.Amigo;
+import com.unir.Usuarios.model.db.Usuario;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,10 +17,15 @@ import java.util.Set;
 public class AmigosService {
 
     @Autowired
+    private UsuariosService usuariosService;
+
+    @Autowired
     private AmigosJpaRepository repository;
 
-    public Amigo createAmigo(String usuario1, String usuario2){
-        Amigo amigo = new Amigo(usuario1, usuario2, "pendiente");
+    public Amigo createAmigo(String usuario1, String usuario2) throws Exception {
+        Usuario user1 = usuariosService.getUsuario("",usuario1,"");
+        Usuario user2 = usuariosService.getUsuario("",usuario2,"");
+        Amigo amigo = new Amigo(user1, user2, "pendiente");
         return repository.save(amigo);
     }
 
@@ -29,40 +35,49 @@ public class AmigosService {
             amigo.setEstado("aceptado");
             repository.save(amigo);
         } else {
-            throw new Exception("Se encontraron mas de una peticion de amigos");
+            throw new Exception("No se encontro peticion de amigo.");
+        }
+        Amigo amigo2 = getAmigo(usuario2, usuario1);
+        if (amigo2==null){
+            amigo2 = new Amigo(amigo.getUsuario2(), amigo.getUsuario1(), "aceptado");
+            repository.save(amigo2);
+        } else {
+            amigo2.setEstado("aceptado");
+            repository.save(amigo2);
         }
     }
 
-    public List<Amigo> getAmigos(String usuario){
+    public List<Amigo> getAmigos(String usuario) throws Exception {
+        Usuario user = usuariosService.getUsuario("",usuario,"");
         Set<Amigo> list = new HashSet<>();
-        repository.findByUsuario1OrUsuario2(usuario, usuario).forEach(amigo -> list.add((Amigo) amigo));
+        repository.findByUsuario1(user).forEach(amigo -> list.add((Amigo) amigo));
         return new ArrayList<>(list);
     }
 
-    public List<Amigo> getAmigos(String usuario, String estado){
+    public List<Amigo> getAmigos(String usuario, String estado) throws Exception {
+        Usuario user = usuariosService.getUsuario("",usuario,"");
         List<Amigo> list = new ArrayList<>();
-        repository.findByUsuario1AndEstado(usuario, estado).forEach(amigo -> list.add((Amigo) amigo));
-        repository.findByUsuario2AndEstado(usuario, estado).forEach(amigo -> list.add((Amigo) amigo));
+        repository.findByUsuario1AndEstado(user, estado).forEach(amigo -> list.add((Amigo) amigo));
         return list;
     }
 
     public Amigo getAmigo(String usuario1, String usuario2) throws Exception {
-        List<Amigo> list = new ArrayList<>();
-        repository.findByUsuario1AndUsuario2(usuario1, usuario2).forEach(amigo -> list.add((Amigo) amigo));
-        repository.findByUsuario1AndUsuario2(usuario2, usuario1).forEach(amigo -> list.add((Amigo) amigo));
-        if(list.size()==1){
-            return list.get(0);
-        } else{
-            throw new Exception("Se encontraron mas de una peticion de amigos");
-        }
+        Usuario user1 = usuariosService.getUsuario("",usuario1,"");
+        Usuario user2 = usuariosService.getUsuario("",usuario2,"");
+        return repository.findByUsuario1AndUsuario2(user1, user2);
     }
 
     public void deleteAmigo(String usuario1, String usuario2) throws Exception {
         Amigo amigo = getAmigo(usuario1, usuario2);
+        Amigo amigo2 = getAmigo(usuario2, usuario1);
+        if(amigo==null && amigo2==null){
+            throw new Exception("No se encontraron amigos entre los usuarios");
+        }
         if(amigo!=null){
             repository.delete(amigo);
-        } else {
-            throw new Exception("No se encontraron amigos entre los usuarios");
+        }
+        if(amigo2!=null){
+            repository.delete(amigo2);
         }
     }
 }
